requirements.txt
openai>=1.51.0
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.2
tenacity>=8.2

google-cloud-translate>=3.15
google-cloud-firestore>=2.17.0
google-cloud-core>=2.4
google-auth>=2.35

grpcio>=1.64
protobuf>=4.25,<6

pydantic>=2.7
